atualização da branch

+++ src/main/java/br/com/gjnv/petshop/dto/EnderecoDto.java
@@ -1,10 +1,12 @@
 package br.com.gjnv.petshop.dto;
 import lombok.Data;
 
+import java.util.UUID;
+
 @Data
 public class EnderecoDto {
 
-    private Long id;
+    private int id;
     private String rua;
     private int numero;
     private String bairro;
